requirements.txt
# Frameworki główne
fastapi==0.104.1
uvicorn==0.24.0
kivy==2.2.1

# Komunikacja i API
requests==2.31.0
openai==1.3.5
python-dotenv==1.0.0

# Obsługa backendu i danych
pydantic==2.5.2
jinja2==3.1.2
python-multipart==0.0.6
itsdangerous==2.1.2

# Serwer ASGI
starlette==0.27.0

# Testy
pytest==7.4.4
httpx==0.27.0
